+++ src/index-web.ts
@@ -11,7 +11,7 @@ export * from './chain/tezos/TezosNodeReader';
 export * from './chain/tezos/TezosNodeWriter';
 export * from './chain/tezos/contracts/BabylonDelegationHelper';
 export * from './chain/tezos/contracts/TCFBakerRegistryHelper';
-export * from './chain/tezos/contracts/TZIPSevenTokenHelper';
+export * from './chain/tezos/contracts/DexterTokenHelper';
 
 export * from './identity/tezos/TezosWalletUtil';
 export * from './identity/tezos/TezosFileWallet';